# backend/docs/swagger.yaml
basePath: /api
definitions:
  dto.IntakeSchedule:
    properties:
      dose_amount:
        type: number
      time_slot:
        allOf:
        - $ref: '#/definitions/shared.TimeSlot'
        enum:
        - morning
        - noon
        - evening
        - night
    required:
    - dose_amount
    - time_slot
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.MedicationCreateRequest:
    properties:
      description:
        minLength: 2
        type: string
      form:
        enum:
        - tablet
        - capsule
        - syrup
        - drop
        - injection
        type: string
      manufacturer:
        type: string
      meal_relation:
        allOf:
        - $ref: '#/definitions/shared.MealRelation'
        enum:
        - before_meal
        - after_meal
        - with_meal
        - irrelevant
      name:
        minLength: 2
        type: string
      pills_per_box:
        minimum: 1
        type: integer
      strength_mg:
        type: integer
    required:
    - form
    - meal_relation
    - name
    - pills_per_box
    - strength_mg
    type: object
  dto.MedicationLogResponse:
    properties:
      id:
        type: string
      planned_dose:
        type: number
      taken:
        type: boolean
      time_slot:
        $ref: '#/definitions/shared.TimeSlot'
      timestamp:
        type: string
      user_medication_id:
        type: string
    type: object
  dto.MedicationResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      form:
        type: string
      id:
        type: string
      manufacturer:
        type: string
      meal_relation:
        $ref: '#/definitions/shared.MealRelation'
      name:
        type: string
      pills_per_box:
        type: integer
      strength_mg:
        type: integer
    type: object
  dto.MedicationUpdateRequest:
    properties:
      description:
        minLength: 2
        type: string
      form:
        enum:
        - tablet
        - capsule
        - syrup
        - drop
        - injection
        type: string
      manufacturer:
        minLength: 2
        type: string
      meal_relation:
        allOf:
        - $ref: '#/definitions/shared.MealRelation'
        enum:
        - before_meal
        - after_meal
        - with_meal
        - irrelevant
      name:
        minLength: 2
        type: string
      strength_mg:
        type: integer
    type: object
  dto.UserCreateRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.UserMedicationCreateRequest:
    properties:
      boxes_owned:
        minimum: 1
        type: integer
      duration_days:
        minimum: 1
        type: integer
      medication_id:
        type: string
      schedules:
        items:
          $ref: '#/definitions/dto.IntakeSchedule'
        minItems: 1
        type: array
    required:
    - boxes_owned
    - duration_days
    - medication_id
    - schedules
    type: object
  dto.UserMedicationResponse:
    properties:
      active:
        type: boolean
      boxes_owned:
        type: integer
      created_at:
        type: string
      id:
        type: string
      medication_id:
        type: string
      schedules:
        items:
          $ref: '#/definitions/dto.IntakeSchedule'
        type: array
      start_at:
        type: string
      user_id:
        type: string
    type: object
  dto.UserMedicationStatsResponse:
    properties:
      daily_consumption:
        type: number
      estimated_days_remaining:
        type: integer
      estimated_end_date:
        type: string
      remaining_pills:
        type: number
      total_pills:
        type: integer
      used_pills:
        type: number
      warning_level:
        description: '"normal", "warning", "critical"'
        type: string
    type: object
  dto.UserMedicationUpdateRequest:
    properties:
      active:
        type: boolean
      boxes_owned:
        minimum: 1
        type: integer
      schedules:
        items:
          $ref: '#/definitions/dto.IntakeSchedule'
        minItems: 1
        type: array
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
  shared.MealRelation:
    enum:
    - before_meal
    - after_meal
    - with_meal
    - irrelevant
    type: string
    x-enum-varnames:
    - MealBefore
    - MealAfter
    - MealWith
    - MealIrregular
  shared.TimeSlot:
    enum:
    - morning
    - noon
    - evening
    - night
    type: string
    x-enum-varnames:
    - Morning
    - Noon
    - Evening
    - Night
info:
  contact: {}
  description: Medication tracking and dose logging API
  termsOfService: http://swagger.io/terms/
  title: DoseLog API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /me:
    get:
      consumes:
      - application/json
      description: Get authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
  /medication-logs/{id}/mark-taken:
    put:
      consumes:
      - application/json
      description: Mark a medication log as taken
      parameters:
      - description: Medication Log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark dose as taken
      tags:
      - medication-logs
  /medication-logs/user-medication/{user_medication_id}:
    get:
      consumes:
      - application/json
      description: Get all logs for a user medication tracking
      parameters:
      - description: User Medication ID
        in: path
        name: user_medication_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MedicationLogResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get medication logs
      tags:
      - medication-logs
  /medications:
    get:
      consumes:
      - application/json
      description: Get paginated list of medications
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MedicationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List medications
      tags:
      - medications
    post:
      consumes:
      - application/json
      description: Add a new medication to the system
      parameters:
      - description: Medication details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MedicationCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MedicationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create medication
      tags:
      - medications
  /medications/{id}:
    get:
      consumes:
      - application/json
      description: Get medication details by ID
      parameters:
      - description: Medication ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MedicationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get medication by ID
      tags:
      - medications
    put:
      consumes:
      - application/json
      description: Update medication details
      parameters:
      - description: Medication ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated medication details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MedicationUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MedicationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update medication
      tags:
      - medications
  /user-medications:
    get:
      consumes:
      - application/json
      description: Get all medication trackings for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserMedicationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user medications
      tags:
      - user-medications
    post:
      consumes:
      - application/json
      description: Create user medication tracking with schedules
      parameters:
      - description: Tracking details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserMedicationCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserMedicationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start medication tracking
      tags:
      - user-medications
  /user-medications/{id}:
    put:
      consumes:
      - application/json
      description: Update user medication tracking details
      parameters:
      - description: User Medication ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated tracking details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserMedicationUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserMedicationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update medication tracking
      tags:
      - user-medications
  /user-medications/{id}/stats:
    get:
      consumes:
      - application/json
      description: Get detailed statistics about medication usage and remaining pills
      parameters:
      - description: User Medication ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserMedicationStatsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get medication statistics
      tags:
      - user-medications
  /user-medications/active:
    get:
      consumes:
      - application/json
      description: Get active medication trackings for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserMedicationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get active user medications
      tags:
      - user-medications
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"